What is tax invoice
A GST invoice, also known as a tax invoice or bill, is a document that a registered business entity issues to its customers for the supply of goods or services. It serves as evidence of the transaction and includes details such as the seller's and buyer's information, itemized description of the goods or services, quantity, price, taxes applicable (including Goods and Services Tax or GST), and any discounts or deductions. The GST invoice is important for tax compliance purposes, as it enables businesses to claim input tax credits and helps ensure proper accounting and reporting of taxes